pyproject.toml
[tool.poetry]
name = "hexlet_code"
version = "0.1.0"
description = ""
authors = ["Data-Wan <ustimenko.work@gmail.com>"]
packages = [
  { include = "calculate_diff" },
]

[tool.poetry.dependencies]
python = "^3.6"
pytest = "^6.2.2"
pytest-cov = "^2.11.1"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
wemake-python-styleguide = "^0.14.1"
pytest = "^6.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gendiff = "calculate_diff.scripts.gendiff:main"